En el kernel del middleware se agregaron los scopes, scope propios de passport los cuales permiten ver los permisos que tendrian un usuario por medio del middleware recien creado llamado role o checkrole, que revisa el rol del usuario y dependiendo de este, dictamina si este usuario puede acceder a los endpoints de las rutas api como post,get,update o delete en el authservices provider definimos los diferentes tipos de permisos en el token que sera enviado al front-end en la response depdendiendo del token que tenga el usuario tendra acceso o no sera permitido el acceso a ese recurso por el error 403 forbidden, en las rutas api se moviendo dentro del grupo middleware para chequear que este autorizado en todo momento al intentar acceder y por su rol antes mencionado, en el front-end modifique lso componentes admin y servicios para enviar el bearrer del usuario conectado que esta guardado en el localstorage, se agregaron unas nuevas rutas para admin y para productos

+++ MascotasFresh-Backend/routes/api.php
@@ -17,17 +17,28 @@ use Illuminate\Support\Facades\Route;
 Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
-/*--------------- Rutas Animales ----------------------------*/
-Route::get('/animales','AnimalController@index');
-Route::post('/agregaranimal', 'AnimalController@store');
-Route::put('/modificaranimal', 'AnimalController@update');
-Route::delete('/eliminaranimal/{id}', 'AnimalController@destroy');
-/*----------------Rutas Usuario ---------------------------------*/
-Route::get('/user', 'UserController@user')->middleware('auth:api');
-//Route::post('/registro', 'UserController@register'); no usaremos el registro hasta que existan algunos usuarios fuera de la veterinaria
-Route::get('/users','UserController@index');
-Route::post('/newuser','UserController@store');
 
-/*----------------------------citas-------------------------------*/
-Route::post('/agregarcita', 'MedicoServicioAnimalController@store');
-Route::get('/serviciosDisponibles','ServicioController@index');
+/*----------------Rutas Middleware---------------------------*/
+Route::middleware(['auth:api','role'])->group(function() {
+    /*------------------------------------ Rutas Animales ----------------------------------------------*/
+    Route::middleware(['scope:Admin,Medico'])->get('/animales','AnimalController@index');
+    Route::middleware(['scope:Admin,Medico'])->post('/agregaranimal', 'AnimalController@store');
+    Route::middleware(['scope:Admin,Medico'])->put('/modificaranimal', 'AnimalController@update');
+    Route::middleware(['scope:Admin,Medico'])->delete('/eliminaranimal/{id}', 'AnimalController@destroy');
+
+    /*----------------------Rutas Usuario ---------------------------------*/
+    //Route::post('/registro', 'UserController@register'); no usaremos el registro hasta que existan algunos usuarios fuera de la veterinaria
+    Route::middleware(['scope:Admin'])->get('/users','UserController@index');
+    Route::get('/user', 'UserController@user')->middleware('auth:api');
+    Route::middleware(['scope:Admin'])->post('/newuser','UserController@store');
+
+    /*--------------------------------------------------citas-------------------------------------------*/
+    Route::middleware(['scope:Admin,Medico'])->post('/agregarcita', 'MedicoServicioAnimalController@store');
+    Route::middleware(['scope:Admin,Medico'])->get('/serviciosDisponibles','ServicioController@index');
+
+
+});
+
+
+
+
